fix(embed): pin transformers<5.0 and state torch_dtype explicitly
The embed and rerank CPU images installed 'transformers>=4.44.2' with no
upper bound, so every rebuild re-resolved to whatever is current, while
Dockerfile.clip.cpu already carries a <5.0 pin. Both now match.

embed_server.py additionally passes torch_dtype=torch.float32 to
AutoModel.from_pretrained instead of inheriting the library default —
the change that actually removes the drift exposure; a moved default
would shift dense CLS+L2 vectors and ReLU-boundary sparse weights
silently. Covered by a route-level test asserting the load kwargs.

Closes #76

+++ src/embed_server.py
@@ -70,7 +70,9 @@ SPARSE_LINEAR_FILE = os.environ.get("SPARSE_LINEAR_FILE", "sparse_linear.pt")
 app = FastAPI(title="vllm-service embed-only", version="1.0")
 
 tokenizer = AutoTokenizer.from_pretrained(MODEL_ID)
-model = AutoModel.from_pretrained(MODEL_ID)
+# Explicit dtype: the float32 default is exactly what a transformers
+# major bump changes, and dense CLS+L2 vectors drift silently if it moves.
+model = AutoModel.from_pretrained(MODEL_ID, torch_dtype=torch.float32)
 model.train(False)
